chat/requirements.txt
# Model Context Protocol
mcp>=1.0.0

# Local LLM inference
llama-cpp-python>=0.2.0

# Async support
asyncio-compat>=0.0.3

# Environment management
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.7.0
click>=8.1.0

# Optional: For enhanced functionality
# streamlit>=1.28.0          # For web interface
# fastapi>=0.104.0           # For API server
# uvicorn>=0.24.0            # For FastAPI server
# requests>=2.31.0           # For HTTP requests
# pydantic>=2.0.0            # For data validation